Refactor event system to centralize event builder
- Remove per-domain event files in favor of a shared defineEvent utility
- Move event definitions into their respective domain index files -
Update exports and imports to use the new event structure
