+++ packages/core/package.json
@@ -15,8 +15,6 @@
     "./site": "./src/site/index.ts",
     "./feed": "./src/feed/index.ts",
     "./parser": "./src/parser/index.ts",
-    "./article/events": "./src/article/events.ts",
-    "./feed/events": "./src/feed/events.ts",
     "./parser/sites": "./src/parser/sites/index.ts",
     "./*": "./src/*.ts"
   },
