--- packages/core/src/site/events.ts
@@ -1,4 +1,0 @@
-// Site events - placeholder
-export namespace Site {
-  export const Events = {};
-}
